Move running view into `frontend::running` module
